refactor(encoding): cleanup package (#2081)
* remove unused reverseBits fcts

* add comment to gnark/icicle backends (cpu vs gpu)

* move EncoderDevice interface to parametrized_encoder file

* move encode/decode.go to encoder.go

* clean up encoder tests into a single file

* consolidate parametrized_encoder

* delete unused rs/params.go file

* cleanup encoding/rs/utils

* fix lint

* address dmanc comment: icicle is gpu/cpu

* make fmt

* fix test

* fix icicle

* docs: address bowen's comments

* fix lint
