--- encoding/rs/decode.go
@@ -1,95 +1,0 @@
-package rs
-
-import (
-	"errors"
-
-	"github.com/Layr-Labs/eigenda/encoding"
-
-	"github.com/consensys/gnark-crypto/ecc/bn254/fr"
-)
-
-// Decode data when some chunks from systematic nodes are lost. It first uses FFT to recover
-// the whole polynomial. Then it extracts only the systematic chunks.
-// It takes a list of available frame, and return the original encoded data
-// storing the evaluation points, since it is where RS is applied. The input frame contains
-// the coefficient of the interpolating polynomina, hence interpolation is needed before
-// recovery.
-// maxInputSize is the upper bound of the original data size. This is needed because
-// the Frames and indices don't encode the length of the original data. If maxInputSize
-// is smaller than the original input size, decoded data will be trimmed to fit the maxInputSize.
-func (e *Encoder) Decode(frames []FrameCoeffs, indices []uint64, maxInputSize uint64, params encoding.EncodingParams) ([]byte, error) {
-	// Get encoder
-	g, err := e.GetRsEncoder(params)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(frames) != len(indices) {
-		return nil, errors.New("number of frames must equal number of indices")
-	}
-
-	// Remove duplicates
-	frameMap := make(map[uint64]FrameCoeffs, len(indices))
-	for i, frameIndex := range indices {
-		_, ok := frameMap[frameIndex]
-		if !ok {
-			frameMap[frameIndex] = frames[i]
-		}
-	}
-
-	numSys := encoding.GetNumSys(maxInputSize, g.ChunkLength)
-	if uint64(len(frameMap)) < numSys {
-		return nil, errors.New("number of frame must be sufficient")
-	}
-
-	samples := make([]*fr.Element, g.NumEvaluations())
-	// copy evals based on frame coeffs into samples
-	for d, f := range frameMap {
-		e, err := GetLeadingCosetIndex(d, g.NumChunks)
-		if err != nil {
-			return nil, err
-		}
-
-		evals, err := g.GetInterpolationPolyEval(f, uint32(e))
-		if err != nil {
-			return nil, err
-		}
-
-		// Some pattern i butterfly swap. Find the leading coset, then increment by number of coset
-		for j := uint64(0); j < g.ChunkLength; j++ {
-			p := j*g.NumChunks + uint64(e)
-			samples[p] = new(fr.Element)
-			samples[p].Set(&evals[j])
-		}
-	}
-
-	reconstructedData := make([]fr.Element, g.NumEvaluations())
-	missingIndices := false
-	for i, s := range samples {
-		if s == nil {
-			missingIndices = true
-			break
-		}
-		reconstructedData[i] = *s
-	}
-
-	if missingIndices {
-		var err error
-		reconstructedData, err = g.Fs.RecoverPolyFromSamples(
-			samples,
-			g.Fs.ZeroPolyViaMultiplication,
-		)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	reconstructedPoly, err := g.Fs.FFT(reconstructedData, true)
-	if err != nil {
-		return nil, err
-	}
-
-	data := ToByteArray(reconstructedPoly, maxInputSize)
-
-	return data, nil
-}
