+++ encoding/rs/encoder.go
@@ -3,8 +3,10 @@ package rs
 import (
 	"errors"
 	"fmt"
+	"log/slog"
 	"math"
 	"sync"
+	"time"
 
 	"github.com/Layr-Labs/eigenda/encoding"
 	"github.com/Layr-Labs/eigenda/encoding/fft"
@@ -18,11 +20,6 @@ type Encoder struct {
 
 	mu                  sync.Mutex
 	ParametrizedEncoder map[encoding.EncodingParams]*ParametrizedEncoder
-}
-
-// Proof device represents a device capable of computing reed-solomon operations.
-type EncoderDevice interface {
-	ExtendPolyEval(coeffs []fr.Element) ([]fr.Element, error)
 }
 
 // NewEncoder creates a new encoder with the given options
@@ -40,9 +37,160 @@ func NewEncoder(config *encoding.Config) (*Encoder, error) {
 	return e, nil
 }
 
-// GetRsEncoder returns a parametrized encoder for the given parameters.
+// just a wrapper to take bytes not Fr Element
+func (g *Encoder) EncodeBytes(inputBytes []byte, params encoding.EncodingParams) ([]FrameCoeffs, []uint32, error) {
+	inputFr, err := ToFrArray(inputBytes)
+	if err != nil {
+		return nil, nil, fmt.Errorf("cannot convert bytes to field elements, %w", err)
+	}
+	return g.Encode(inputFr, params)
+}
+
+// Encode function takes input in unit of Fr Element and creates a list of FramesCoeffs,
+// which each contain a list of multireveal interpolating polynomial coefficients.
+// A slice of uint32 is also returned, which corresponds to which leading coset
+// root of unity the frame is proving against. This can be deduced from a frame's index.
+func (g *Encoder) Encode(inputFr []fr.Element, params encoding.EncodingParams) ([]FrameCoeffs, []uint32, error) {
+	start := time.Now()
+	intermediate := time.Now()
+
+	// Get RS encoder from params
+	encoder, err := g.getRsEncoder(params)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	pdCoeffs, err := encoder.padPolyEval(inputFr)
+	if err != nil {
+		return nil, nil, err
+	}
+	paddingDuration := time.Since(intermediate)
+
+	intermediate = time.Now()
+
+	polyEvals, err := encoder.RSEncoderComputer.ExtendPolyEval(pdCoeffs)
+	if err != nil {
+		return nil, nil, fmt.Errorf("reed-solomon extend poly evals, %w", err)
+	}
+	extensionDuration := time.Since(intermediate)
+
+	intermediate = time.Now()
+
+	// create Frames to group relevant info
+	frames, indices, err := encoder.makeFrames(polyEvals)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	framesDuration := time.Since(intermediate)
+
+	// TODO(samlaf): use an injected logger instead.
+	slog.Info("RSEncode details",
+		"input_size_bytes", len(inputFr)*encoding.BYTES_PER_SYMBOL,
+		"num_chunks", encoder.Params.NumChunks,
+		"chunk_length", encoder.Params.ChunkLength,
+		"padding_duration", paddingDuration,
+		"extension_duration", extensionDuration,
+		"frames_duration", framesDuration,
+		"total_duration", time.Since(start))
+
+	return frames, indices, nil
+}
+
+// Decode data when some chunks from systematic nodes are lost. This function implements
+// https://ethresear.ch/t/reed-solomon-erasure-code-recovery-in-n-log-2-n-time-with-ffts/3039
+//
+// It first uses FFT to recover the whole polynomial. Then it extracts only the systematic chunks.
+// It takes a list of available frame, and return the original encoded data
+// storing the evaluation points, since it is where RS is applied. The input frame contains
+// the coefficient of the interpolating polynomina, hence interpolation is needed before
+// recovery.
+//
+// maxInputSize is the upper bound of the original data size. This is needed because
+// the Frames and indices don't encode the length of the original data. If maxInputSize
+// is smaller than the original input size, decoded data will be trimmed to fit the maxInputSize.
+func (e *Encoder) Decode(
+	frames []FrameCoeffs, indices []uint64, maxInputSize uint64, params encoding.EncodingParams,
+) ([]byte, error) {
+	// Get encoder
+	g, err := e.getRsEncoder(params)
+	if err != nil {
+		return nil, err
+	}
+
+	if len(frames) != len(indices) {
+		return nil, errors.New("number of frames must equal number of indices")
+	}
+
+	// Remove duplicates
+	frameMap := make(map[uint64]FrameCoeffs, len(indices))
+	for i, frameIndex := range indices {
+		_, ok := frameMap[frameIndex]
+		if !ok {
+			frameMap[frameIndex] = frames[i]
+		}
+	}
+
+	numSys := encoding.GetNumSys(maxInputSize, g.Params.ChunkLength)
+	if uint64(len(frameMap)) < numSys {
+		return nil, errors.New("number of frame must be sufficient")
+	}
+
+	samples := make([]*fr.Element, g.Params.NumEvaluations())
+	// copy evals based on frame coeffs into samples
+	for d, f := range frameMap {
+		e, err := GetLeadingCosetIndex(d, g.Params.NumChunks)
+		if err != nil {
+			return nil, err
+		}
+
+		evals, err := g.getInterpolationPolyEval(f, e)
+		if err != nil {
+			return nil, err
+		}
+
+		// Some pattern i butterfly swap. Find the leading coset, then increment by number of coset
+		for j := uint64(0); j < g.Params.ChunkLength; j++ {
+			p := j*g.Params.NumChunks + uint64(e)
+			samples[p] = new(fr.Element)
+			samples[p].Set(&evals[j])
+		}
+	}
+
+	reconstructedData := make([]fr.Element, g.Params.NumEvaluations())
+	missingIndices := false
+	for i, s := range samples {
+		if s == nil {
+			missingIndices = true
+			break
+		}
+		reconstructedData[i] = *s
+	}
+
+	if missingIndices {
+		var err error
+		reconstructedData, err = g.Fs.RecoverPolyFromSamples(
+			samples,
+			g.Fs.ZeroPolyViaMultiplication,
+		)
+		if err != nil {
+			return nil, fmt.Errorf("recover polynomial from samples: %w", err)
+		}
+	}
+
+	reconstructedPoly, err := g.Fs.FFT(reconstructedData, true)
+	if err != nil {
+		return nil, fmt.Errorf("inverse fft on reconstructed data: %w", err)
+	}
+
+	data := ToByteArray(reconstructedPoly, maxInputSize)
+
+	return data, nil
+}
+
+// getRsEncoder returns a parametrized encoder for the given parameters.
 // It caches the encoder for reuse.
-func (g *Encoder) GetRsEncoder(params encoding.EncodingParams) (*ParametrizedEncoder, error) {
+func (g *Encoder) getRsEncoder(params encoding.EncodingParams) (*ParametrizedEncoder, error) {
 	g.mu.Lock()
 	defer g.mu.Unlock()
 	enc, ok := g.ParametrizedEncoder[params]
@@ -72,7 +220,7 @@ func (e *Encoder) newEncoder(params encoding.EncodingParams) (*ParametrizedEncod
 		return nil, err
 	}
 
-	fs := e.CreateFFTSettings(params)
+	fs := e.createFFTSettings(params)
 
 	switch e.Config.BackendType {
 	case encoding.GnarkBackend:
@@ -84,7 +232,7 @@ func (e *Encoder) newEncoder(params encoding.EncodingParams) (*ParametrizedEncod
 	}
 }
 
-func (e *Encoder) CreateFFTSettings(params encoding.EncodingParams) *fft.FFTSettings {
+func (e *Encoder) createFFTSettings(params encoding.EncodingParams) *fft.FFTSettings {
 	n := uint8(math.Log2(float64(params.NumEvaluations())))
 	return fft.NewFFTSettings(n)
 }
@@ -96,7 +244,7 @@ func (e *Encoder) createGnarkBackendEncoder(params encoding.EncodingParams, fs *
 
 	return &ParametrizedEncoder{
 		Config:            e.Config,
-		EncodingParams:    params,
+		Params:            params,
 		Fs:                fs,
 		RSEncoderComputer: &gnarkencoder.RsGnarkBackend{Fs: fs},
 	}, nil
