--- encoding/rs/encoder_fuzz_test.go
@@ -1,38 +1,0 @@
-package rs_test
-
-import (
-	"testing"
-
-	"github.com/Layr-Labs/eigenda/encoding"
-	"github.com/Layr-Labs/eigenda/encoding/rs"
-	"github.com/stretchr/testify/assert"
-)
-
-func FuzzOnlySystematic(f *testing.F) {
-
-	f.Add(GETTYSBURG_ADDRESS_BYTES)
-	f.Fuzz(func(t *testing.T, input []byte) {
-
-		params := encoding.ParamsFromSysPar(10, 3, uint64(len(input)))
-		cfg := encoding.DefaultConfig()
-		enc, err := rs.NewEncoder(cfg)
-		if err != nil {
-			t.Errorf("Error making rs: %q", err)
-		}
-
-		//encode the data
-		frames, _, err := enc.EncodeBytes(input, params)
-		if err != nil {
-			t.Errorf("Error Encoding:\n Data:\n %q \n Err: %q", input, err)
-		}
-
-		//sample the correct systematic Frames
-		samples, indices := sampleFrames(frames, uint64(len(frames)))
-
-		data, err := enc.Decode(samples, indices, uint64(len(input)), params)
-		if err != nil {
-			t.Errorf("Error Decoding:\n Data:\n %q \n Err: %q", input, err)
-		}
-		assert.Equal(t, input, data, "Input data was not equal to the decoded data")
-	})
-}
