+++ encoding/rs/utils.go
@@ -90,21 +90,6 @@ func ToByteArray(dataFr []fr.Element, maxDataSize uint64) []byte {
 	return data
 }
 
-func GetNumElement(dataLen uint64, CS int) uint64 {
-	numEle := int(math.Ceil(float64(dataLen) / float64(CS)))
-	return uint64(numEle)
-}
-
-// helper function
-func RoundUpDivision(a, b uint64) uint64 {
-	return uint64(math.Ceil(float64(a) / float64(b)))
-}
-
-func NextPowerOf2(d uint64) uint64 {
-	nextPower := math.Ceil(math.Log2(float64(d)))
-	return uint64(math.Pow(2.0, nextPower))
-}
-
 // This function is used by user to get the leading coset for a frame, where i is frame index
 func GetLeadingCosetIndex(i uint64, numChunks uint64) (uint32, error) {
